refactor: refactor interchain test
